backend-archive/python-llm/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
mangum>=0.17.0

